.claude/worktrees/xenodochial-ritchie-3e318c/backend/requirements.txt
# Web Framework
fastapi==0.115.1
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
# Prebuilt wheels on Python 3.13 (0.21.x often builds from source)
watchfiles>=1.0.0

# Database
# SQLAlchemy 2.0.23 crashes on import with Python 3.13 (typing / TypingOnly)
sqlalchemy[asyncio]>=2.0.36,<2.1.0
# asyncpg 0.29 does not build on Python 3.13 (C API)
asyncpg>=0.30.0
# 2.9.9 has no cp313 wheel on macOS/arm64; falls back to source + pg_config
psycopg2-binary>=2.9.11
alembic==1.13.0
pgvector>=0.3.0

# Data Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0
# 2.1.0 yanked on PyPI
email-validator>=2.1.1,<3.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Job Queue
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx>=0.28.1
aiohttp>=3.9.0

# AI/ML - LLM Providers (pinned so pip resolves fast)
google-genai
anthropic>=0.40.0
openai==1.12.0

# MCP - Model Context Protocol (server + client)
mcp==1.2.0
# mcp pulls sse-starlette 3.x which requires starlette>=0.49; FastAPI 0.115.x needs starlette<0.39
sse-starlette>=2.1.3,<3

# Google Calendar API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Microsoft Outlook Calendar (MSAL + Graph API via httpx)
msal>=1.28.0

# Telegram
python-telegram-bot>=21.0

# Utilities
python-dotenv>=1.1.1,<2.0.0
pydantic-extra-types>=2.1.0
bleach>=6.1.0                  # XSS sanitization

# Logging
python-json-logger==2.0.7

# Resilience — retry + circuit breaker
tenacity==8.5.0
pybreaker==1.2.0

# Conversation state machine
transitions==0.9.2

# Observability — OpenTelemetry
opentelemetry-sdk>=1.24.0,<2.0.0
opentelemetry-exporter-otlp-proto-http>=1.24.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-instrumentation-sqlalchemy>=0.44b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov>=4.1.0               # Test coverage
aiosqlite>=0.19.0               # SQLite async for testing

# Date/Time utilities
python-dateutil==2.8.2
pytz==2023.3

# Error monitoring
sentry-sdk[fastapi]>=2.0.0
